product_service/requirements.txt
Flask==2.3.2
psycopg2-binary==2.9.3